Build stm-containers instead of the one from nixpkgs.
As noted in
https://github.com/noteed/start-servant/pull/11#discussion_r693383662,
the package is marked as broken in nixpkgs. So we'll just pin it (and
some of its unavailable dependencies) and build it ourselves.

This has been done in nix/stm-containers.nix.

@noteed
This fix should suffice for using stm-containers.

I had to also add `containers` to the cabal file since it was needed
from your changes in #11; maybe you forgot to stage that change?