Cleanup 1: Remove redundant imports, orphan instances.
I've removed all imports that were redundant as the first phase; this
gives me a clear idea of what imports are needed and where.

The next phase is to now start with removing the default Prelude and
replacing it with Protolude.

This will be a more invasive change; and will also involving replacing
String with Text.

I've also added .envrc to gitignore, since we somewhat agree on not
committing this due to secrets in the envrc files.