Runtime: with modes.
Introducing the runtime and configuration. An application should start
with a configuration; and the attempt to construct a Runtime in a
suitable mode (possibly indicated using arguments); and start things
up.

Modes can be Stm or Postgres; and the right "Handler" will be used in
the runtime based on the mode.

This is again type-safe; and mixups will be easily avoided.

This gives us a pretty good starting ground for being able to
experiment with multiple storage systems underneath.

Following this; we'll start constructing generalised ways for
manipulating stored data; along with ACL.