Boot process with modes.
This is an example of how to boot the application in differing modes.

Currently, we just support the STM mode; but we can trivially add
support for both STM and postgres mode in the near future.

My next steps would be to add abstractions over the runtime and
storage to provide mode-agnostic storages.