Basics of abstract storage.
The goal of these typeclasses is to abstract /how/ data is stored; and
only rely on some environment/monad @m@; for which we can implement
these instances.

The core functions here are just
- dbUpdate
- dbSelect

And these functions can be implemented more concretely with the
definition of the @m@ in which they operate.