Examples of how the abstract storage actually wires in.
The first implementations (along with some stubs) are for users; we'll
now see how these classes help us perform these operations per `m`;
however; we'll leave the postgres implementations for later. But it
should be clear that the choice of `m` determines the more intimate
execution details, not the overall interface.