Added a TH level function that lets us get module information.
@noteed this will help us with https://github.com/noteed/start-servant/issues/2

Basically, each page can just generate this value at runtime; and we
can ofcourse use the root repo. URL to also link these pages to GH.

How this works with GH branches etc. remains to be seen; but I believe
we can also compute that if we have the git environment available to
us: GH uses the branch name and the commit hashes to point to URL
"blobs".

But this will get us there.