Fix logging: properly format sentences.
@noteed I've added custom loggging functions in Prelude; and these
handle the sentence termination for us.

This lets us avoid these errors and also avoid boilerplate when
wanting to properly terminate sentences while logging.

Imagine, if we were to want to log errors:

Normally, we'd have to do something like:

```
case eResult of
  Left err -> L.error $ "Encountered an error! Err: " <> show err <>
  "."

```

Which is prone to being forgotten. The custom logging functions handle
this for us.

There are 2 functions, sentence and ellipsis which terminate the
sentences in different ways.

Conversation: https://github.com/noteed/start-servant/pull/14#discussion_r704687526