Wire it all in: fire up the server in the `New` mode!
We've been clever here: the only information of /what/ the actual `m`
we're running in is only encoded in the main module.

This lets us be totally agnostic of the `m` we're running in, and only
reason about its properties.

The big win? We can operate within Stm storage modes; or within a
Postgres storage modes with /exactly/ the same handlers.