Format StmDatabase module.
@noteed I've formatted this module with Brittany; which I personally
prefer.
But Brittany is known to be pretty opinionated; however, I've seen it
consistently achieves the best results. Brittany's opionated nature
also means that the same piece of code results in the same block of
code no matter who the programmer might be.

Let me know if this is fine for you; or we need to look into other
formatters.