Extra constraints, and provide the Todo page with the best auth.
We now wire it all up; attempting to authorize with RW if possible,
with a fallback to RO for each todo list.

The markup of each Todo-list isn't done, but with this mechanism, if
the user has RW permissions on the list, we display the list with RW;
otherwise we fallback to RO.

Next; we can provide separate newtype wrappers over todolists to be
able to modify the state of an item for the RW display.

@noteed; this should be pretty handy to provide the same routes for
varying permissions and displaying different views depending on the permissions.