Permissions with todo-lists.
The user can now, using our ACL machinery, view todo-lists with the
most permissive mode as allowed by the ACL of the user.

In the next commit, we introduce 2 todo lists: one with RO access to
the user alice, and another with RW; and the modes can then be seen in action.