Rename list and item fields for easy lenses.
Lenses will prove useful for updating items in place in the following
section and hence we rename the fields.

`makeLenses` is optimal and easy to use; but the one condition it has
is prepending field accessors with `_`; which is only a small
compromise (IMHO) since lenses help a lot with traversing through
nested structures.