Storage helper functions.
When we do a dbUpdate; the function should return a list of IDs that
were affected by the update.

We now define the common use-case helper functions that can act on the
returned IDs (to retrieved the actual stored resources that were
affected, for example).

There are variants of this function:
- One that just gets all affected.
- One that gets the first one (when we're sure we've only affected a
single resource)
- One that gets the first one, and errors out if there are none.