Full flow: marking a TodoItem to a different status.
Mark items in varying states.

This also takes care of the ACL of the user.

For example, with the user `alice`; she can edit items of list TL-1
but not of TL-0.

Eg. try http://localhost:7249/private/user/todos/TL-0/item/item0/mark?newState=Done

which will lead to a permission denied error, as we'd expect.