Editing TodoItems.
Editing todo-items was also simple to achieve.
All we needed was a FromForm instance; and voila.

Things are now getting in place: we have a nice "DSL" to express
authentication, and authorization.

@noteed this completes the TodoItem level stubs.
I'll now move on to creating new lists, and attributing the right
permissions on them via our ACL mechanism.

Once that is done, we can move into the design phase.