Use of ViewMode.
As mentioned in the parent commit; here we see ViewModes in action.

The one downside of this approach has been the introduction of
orphans. However, this is not an outcome of the design per se, but the
fact that we've  not yet divided our "Types" modules into smaller
sub-modules yet, and these instances can then be moved to these
smaller modules.

However, I feel pretty torn on the Orphan instance limitations of any
typeclass based system. It can really become a nuissance at times, and
whether these warnings have any weight depends on the use case.

Eg. for a library, orphans are a detriment. But are they the same for
applications?

@noteed thoughts?