Get rid of the whole Prelude mess.
Cabal mixins are half baked and lead to very unreasonable behaviour.
At this time, its just better to get rid of an implicit prelude and
explicitly import Protolude wherever needed.

Additionally, I've added a separate Logging module which isn't called
Prelude to ensure the names don't conflict: we'd like to use the same
helpers for logging in the prototype-hs-example for logging.

@thusc this was a bit of a waste of time; my apologies for that.