Add a logging module for multiple-logger logging.
The new logging module, as we discussed, @thusc, provides
functionality to log over multiple loggers.

This, is, from what I can gather, fully supportive of our
requirements.

Additionally, I've kept convenience in mind in keeping with having a
typeclass that we needn't care about when dealing with logging in some `m`.